Moves data class JobCompletionData to an inner class and makes it private, and fixes any problems associated with that move
